fix: imports activity with correct date
Previously the input [type="date"] element was converting the resulting value (
a date at midnight) to my local timezone (-6 hours), so the value was being imported
as the previous day.

My quick fix is to use the [type="datetime-local"].  This element isn't widely supported,
so I'll probably need to create my own date time component for the form?